% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGA.R
\name{readGenericTCGA}
\alias{readGenericTCGA}
\title{Parse generic Level 3 TCGA data files}
\usage{
readGenericTCGA(uri)
}
\arguments{
\item{uri}{a url or filename of a TCGA data matrix}
}
\value{
A list with elements \code{val}, a tbl_df and
\code{header} with the second row of the header as a
character vector.
}
\description{
The TCGA Level 3 data files seem to all follow a similar
pattern with the first row being the sample names and the
second row having value descriptions ("Beta_value",
"Expression", etc.). This does the first-level parsing
to give back the "data" as a data.frame and the first
row as column names. The second row is returned as the
second item in the list.
}
\note{
This function is not meant to be called by an
end-user, but it is exported for the purposes of extensibility
by a developer.
}

