% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGA.R
\name{splitTCGAIds}
\alias{splitTCGAIds}
\title{Split TCGA IDs to pieces}
\usage{
splitTCGAIds(ids)
}
\value{
A data frame with the portions of the TCGA ID
split into named columns.
}
\description{
Split TCGA IDs to pieces
}

