% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGA.R
\name{write.afm}
\alias{write.afm}
\title{Write a matrix out as an Annotated Feature Matrix file}
\usage{
write.afm(mat, fname, append = FALSE, ...)
}
\arguments{
\item{mat}{an R matrix}

\item{fname}{the filename to which to write the resulting
afm file}

\item{append}{A logical, whether or not to append the matrix
to a file or overwrite the file. Note that columns should
match the existing file if append=TRUE.}
}
\description{
The Annotated Feature Matrix (.afm) format is described in
detail at \url{https://code.google.com/p/rf-ace/wiki/Manual}.
Basically, though, features are in rows and samples/observations
are in columns. The rows are annotated with a data type that can
include:
\itemize {
  \item{B} for binary data
  \item{C} for categorical data (string)
  \item{N} for numeric data
}
}
\details{
The column names are required as specified in
\url{https://code.google.com/p/rf-ace/wiki/Manual}. The row
names are also required and will be written out with the
appropriate prefix (B, N, C) based on the type of the matrix
to be written.
}
\examples{
mat = matrix(rnorm(25),nc=5)
colnames(mat) = rownames(mat) = LETTERS[1:5]
tmpfile = tempfile()
write.afm(mat,tmpfile)
readLines(tmpfile)
}

